requirements.txt
# MalCapture Defender - Requirements

# Core Dependencies
psutil>=5.9.5              # Process and system monitoring
pywin32>=305               # Windows API access
watchdog>=3.0.0            # File system monitoring
scapy>=2.5.0               # Network packet analysis

# GUI Framework
PyQt5>=5.15.9              # GUI framework
pyqtgraph>=0.13.3          # Real-time plotting

# Data Processing
pandas>=2.0.0              # Data manipulation
numpy>=1.24.0              # Numerical computations

# Logging and Configuration
PyYAML>=6.0                # Configuration file handling
colorlog>=6.7.0            # Colored logging

# API Hooking (Windows)
pyinstaller>=5.13.0        # For building executable
pefile>=2023.2.7           # PE file analysis

# Utilities
python-dateutil>=2.8.2     # Date/time utilities
requests>=2.31.0           # HTTP requests (for updates)
packaging>=23.1            # Version parsing

# Testing (Development)
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reporting
pytest-mock>=3.11.1        # Mocking support

# Security
cryptography>=41.0.0       # Encryption utilities

# Optional: Machine Learning for Behavioral Analysis
scikit-learn>=1.3.0        # ML algorithms
joblib>=1.3.0              # Model persistence
